# zstd: Optimize seqdeq amd64 asm