@klauspost 2022-07-04
@greatroar I can confirm your numbers. Some rather noticeable regressions, but with this I can port some of the s2 memcopy for an even bigger speedup and fixing some of the regression cases.

Thanks for the contribution!